Phân quyền quản trị người dùng cho các user là admin và super admin cũng như quyền của các role trong hệ thống

+++ app/Http/Controllers/Admin/UserController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers\Admin;
+
 use App\Http\Controllers\Controller;
 use App\Models\Faculty;
 use Illuminate\Http\Request;
@@ -19,16 +20,22 @@ class UserController extends Controller
         if ($request->has('item-per-page')) $item_per_page = $request->query('item-per-page');
 
         $query_condition = [
-            'isSystemAccount' => 0
+            'isSystemAccount' => 0,
+            ['id', '!=', Auth::user()['id']]
         ];
 
         if (!Auth::user()['isAdmin']) $query_condition['faculty_id'] = $khoa['id'];
 
         $users = User::where($query_condition)->paginate($item_per_page);
 
-        foreach ($users as $user) {
-            $user['faculty'] = $user->faculty;
-            $user['roles'] = $user->roles;
+        $auth_users_role_level = Auth::user()->roles->max('level');
+
+        foreach ($users as $key => $user) {
+            if ($user->roles->max('level') < $auth_users_role_level) $users->forget($key);
+            else {
+                $user['faculty'] = $user->faculty;
+                $user['roles'] = $user->roles;
+            }
         }
 
         return view('server.pages.user.index', [
@@ -40,7 +47,9 @@ class UserController extends Controller
     public function create(Request $request, $khoa)
     {
         $faculty_list = Faculty::all();
-        $roles = Roles::all();
+        $max_user_level = Auth::user()->roles->max('level');
+
+        $roles = Roles::where([['level', '>',  $max_user_level]])->get();
 
         return view('server.pages.user.create', [
             'khoa' => $khoa,
@@ -53,6 +62,13 @@ class UserController extends Controller
     {
         // Tạo mới user
         $user = new User($request->input());
+
+        // dừng lại nếu tạo user có khoa khác với khoa của user đang login trừ admin
+        abort_if(!Auth::user()['isAdmin'] && $user['faculty_id'] != Auth::user()['faculty_id'], 403);
+
+        // Luu y: chuyen vao policy khu lam toi chuc nang nay
+        // dừng lại nếu user muốn tạo có quyền cao hơn hoặc bằng mình
+        abort_if(Auth::user()->roles->max('level') >= $user->roles->max('level'), 403);
 
         // Mã hóa password
         $user->password = bcrypt($request->password);
@@ -85,14 +101,25 @@ class UserController extends Controller
         // Ngưng nếu ko tìm thấy user
         abort_if(!$user, 404);
 
+        // dừng lại nếu cập nhật user có khoa khác với khoa của user đang login trừ admin
+        abort_if(!Auth::user()['isAdmin'] && $user['faculty_id'] != Auth::user()['faculty_id'], 403);
+
+        // Luu y: chuyen vao policy khu lam toi chuc nang nay
+        // dừng lại nếu user muốn chỉnh sửa có quyền cao hơn hoặc bằng mình trừ chính mình
+        abort_if(Auth::user()->roles->max('level') >= $user->roles->max('level') && Auth::user()['id'] != $user['id'], 403);
+
         // lấy danh sách role của user
         $user['roles_list'] = $user->roles->pluck('id')->toArray();
 
         // Lấy tất cả các khoa
         $faculty_list = Faculty::all();
 
-        // Lấy tất cả role
-        $roles = Roles::all();
+        // Lấy tất cả role level thấp hơn role của mình
+        $max_user_level = Auth::user()->roles->max('level');
+
+        $roles = Roles::where([['level', '>',  $max_user_level]])->get();
+
+        // dd($roles);
 
         return view('server.pages.user.edit', [
             'khoa' => $khoa,
@@ -110,14 +137,24 @@ class UserController extends Controller
         // dừng nếu user không tồn tại
         abort_if(!$user, 404);
 
+        // dừng lại nếu xóa user có khoa khác với khoa của user đang login trừ admin
+        abort_if(!Auth::user()['isAdmin'] && $user['faculty_id'] != Auth::user()['faculty_id'], 403);
+
+        // Luu y: chuyen vao policy khu lam toi chuc nang nay
+        // dừng lại nếu user muốn chỉnh sửa có quyền cao hơn hoặc bằng mình trừ chính mình
+        abort_if(Auth::user()->roles->max('level') >= $user->roles->max('level') && Auth::user()['id'] != $user['id'], 403);
+
         // Mã hóa password
-        if ($request->filled('password')) $request->merge(['password' => bcrypt($request->password)]);
+        if ($request->filled('password')) $request->merge(['password' => bcrypt($request['password'])]);
+        else $request->offsetUnset('password');
 
         // chuyển status sang dạng 1, 0
         if ($request->filled('status')) $request->merge(['status' => $request['status'] == "on" ? 1 : 0]);
+        else $request->offsetUnset('status');
 
         // chuyển isAdmin sang dạng 1, 0
         if ($request->filled('isAdmin')) $request->merge(['isAdmin' => $request['isAdmin'] == "on" ? 1 : 0]);
+        else $request->offsetUnset('isAdmin');
 
         // upload ảnh đại diện
         if ($request->file('avatar') != null) $request->merge(['avatar' => upload_file($request->file('avatar'), 'dist/upload/image/3/users')]);
@@ -152,6 +189,13 @@ class UserController extends Controller
 
         // dừng nếu user không tồn tại
         abort_if(!$user, 404);
+
+        // dừng lại nếu cập nhật user có khoa khác với khoa của user đang login trừ admin
+        abort_if(!Auth::user()['isAdmin'] && $user['faculty_id'] != Auth::user()['faculty_id'], 403);
+
+        // Luu y: chuyen vao policy khu lam toi chuc nang nay
+        // dừng lại nếu user muốn xóa có quyền cao hơn hoặc bằng mình trừ chính mình
+        abort_if(Auth::user()->roles->max('level') >= $user->roles->max('level') && Auth::user()['id'] != $user['id'], 403);
 
         // Gỡ toàn bộ role của user ra
         $user->roles()->detach();
